+++ package.json
@@ -12,7 +12,7 @@
     "dev": "pnpm run --parallel --recursive dev",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "build-storybook": "build-storybook",
-    "storybook": "pnpm run --parallel --recursive storybook:run",
+    "storybook": "pnpm run --parallel --recursive storybook:dev",
     "storybook:build": "turbo run storybook:build",
     "migrations": "cd apps/builder && pnpm migrations --dev",
     "build-figma-tokens": "cd packages/design-system && pnpm build-figma-tokens",
@@ -26,12 +26,12 @@
     "pre-commit": "./node_modules/.bin/nano-staged"
   },
   "devDependencies": {
-    "@storybook/react": "^6.5.16",
     "@types/css-tree": "^2.3.1",
     "@webstudio-is/eslint-config-custom": "workspace:^",
     "nano-staged": "^0.8.0",
     "prettier": "2.8.7",
     "simple-git-hooks": "^2.8.1",
+    "storybook": "^7.0.23",
     "tsx": "^3.12.6",
     "turbo": "^1.8.6",
     "typescript": "5.1.3"
