+++ packages/css-engine/.storybook/main.ts
@@ -1,0 +1,1 @@
+export { default } from "@webstudio-is/storybook-config";
