--- packages/image/.storybook/main.cjs
@@ -1,1 +1,0 @@
-module.exports = require("@webstudio-is/storybook-config");
