+++ packages/sdk-components-react/package.json
@@ -39,8 +39,8 @@
     "dts": "tsc --project tsconfig.dts.json",
     "typecheck": "tsc --noEmit --emitDeclarationOnly false",
     "checks": "pnpm typecheck",
-    "storybook:run": "start-storybook -p 6006",
-    "storybook:build": "build-storybook"
+    "storybook:dev": "storybook dev -p 6006",
+    "storybook:build": "storybook build"
   },
   "peerDependencies": {
     "react": "^18.2.0",
@@ -57,7 +57,7 @@
     "shallow-equal": "^3.1.0"
   },
   "devDependencies": {
-    "@storybook/react": "^6.5.16",
+    "@storybook/react": "^7.0.23",
     "@types/react": "^18.0.35",
     "@types/react-dom": "^18.0.11",
     "@webstudio-is/scripts": "workspace:^",
