+++ packages/storybook-config/package.json
@@ -3,14 +3,17 @@
   "private": true,
   "version": "0.0.0",
   "license": "MIT",
+  "main": "index.ts",
   "dependencies": {
     "@fontsource/inter": "^4.5.15",
     "@fontsource/manrope": "^4.5.13",
     "@fontsource/roboto-mono": "^4.5.10",
-    "@storybook/addon-essentials": "^6.5.16",
-    "@storybook/addon-links": "^6.5.16",
-    "@storybook/builder-vite": "^0.4.2",
+    "@storybook/addon-essentials": "^7.0.23",
+    "@storybook/addon-links": "^7.0.23",
+    "@storybook/react": "^7.0.23",
+    "@storybook/react-vite": "^7.0.23",
     "@vitejs/plugin-react": "^3.1.0",
+    "storybook": "^7.0.23",
     "vite": "^4.1.4"
   }
 }
